requirements.txt
PyQt6
paho-mqtt
PyYAML
